libparse: change implicit sign channge to explicit.

+++ libparse/clk_wharton.c
@@ -122,7 +122,8 @@ inp_wharton_400a(
 {
 	unsigned int rtc;
 	
-	parseprintf(DD_PARSE, ("inp_wharton_400a(0x%lx, 0x%x, ...)\n", (long)parseio, ch));
+	parseprintf(DD_PARSE, ("inp_wharton_400a(0x%lx, 0x%x, ...)\n",
+                    (unsigned long)parseio, (unsigned)ch));
 	
 	switch (ch)
 	{
